Project Initialization - 6
Authentication route has been established. All necessary business logic has been implemented for the register post. Tests have been verified for the authRoute. Middleware to validate the input data is also designed with TDD, but does not contain abstracted testing. The testing of the combination of the authRoute and the middleware is done at the same time. Error codes have been properly assigned. Continued implementation of env variables. Logic to handle tokens has also been implemented, but remains to be tested with the login endpoint. The validateUser middleware will also be tested with the implementation of the login endpoint.